move spearate functions of user interactions into a class
